// src/app/models/academics/sequence-status-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * SequenceStatusView
 */
export interface SequenceStatusView { 
    /**
     * Gets or Sets IsDeleted
     */
    isActive?: number;
    id: string;
}