// src/app/models/academics/school-sequence-viewmodel.ts
import { SequenceListView } from './sequence-list-view';

/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * SchoolSequenceViewModel
 */
export interface SchoolSequenceViewModel {
    /**
     * Gets or Sets SequenceCategoryType_Id
     */
    sequenceCategoryTypeId?: number;
    /**
     * Gets or Sets SequenceParameterIds
     */
    sequenceParameterIds?: Array<SequenceListView>;
    /**
     * Gets or Sets Starting_Sequence
     */
    startingSequence?: number;
    /**
 * Gets or Sets schoolId
 */
    schoolId?: string;

}